Moved Skeletons to another folder
